fix(bms系统数据类型): 修改soc数据类型

+++ src/main/java/com/xiliulou/electricity/dto/bms/BatteryInfoDto.java
@@ -23,7 +23,7 @@ public class BatteryInfoDto {
     /**
      * soc
      */
-    private Integer soc;
+    private Double soc;
     /**
      * 电池容量
      */
